// Backend/src/constants/typeTransaction.enum.ts
export enum TYPE_TRANSACTION{
  GASTE = 0, // PARA CUANDO UN CLIENTE GASTE EN UN VIDEO
	RECIBO_CLIENTE= 1,  // PARA CUANDO UN CLIENTE RECARGUE EN LA EMPRESA (ROCKOBITS)
	TRANSFERI = 2, // PARA CUANDO UNA EMPRESA TRANSFIERE A UN CLIENTE
	COMPRE = 3,    // PARA CUANDO UNA EMPRESA COMPRA ROCKOBITS EN LA PLATAFORMA CON STRIPE.
	RECIBO_EMPRESA = 4, // PARA CUANDO UNA EMPRESA RECIBE ROCKOBITS POR FACTOR EXTERNO
	ENVIA = 5, //PARA CUANDO UN SUPER ADMIN LE RECARGUE ROCKOBITS A UNA EMPRESA  
	TRANSFERI_EMPLEADO = 6, //PARA CUANDO UNA EMPRESA LE TRANSFIERA A UN EMPLEADO
	RECIBO_EMPLEADO = 7, //PARA CUANDO UN EMPLEADO RECIBA DE LA EMPRESA 
}